SilverAI-Assignment-AI-Engineering-main/app/requirements.txt
streamlit
pypdf
pdfplumber
lightrag-hku
supabase
openai
google-generativeai
python-dotenv
tqdm
numpy
torch>=2.2.0
transformers>=4.43.0
datasets
einops>=0.8.0
psycopg2-binary
sentence-transformers
nest_asyncio
asyncio
python-multipart
